requirements.txt
openai-whisper==20231117
pyannote.audio
torch==2.1.2
pydub